# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transport/auth/login:
    post:
      operationId: transportLogin
      tags: [transport-auth]
      summary: Login to transport platform
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransportLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportLoginResponse"
        "401":
          description: Invalid credentials
  /transport/auth/me:
    get:
      operationId: transportMe
      tags: [transport-auth]
      summary: Get current transport user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportUser"
        "401":
          description: Not authenticated

  # ─── TENANT MANAGEMENT (superadmin) ────────────────────────────────────
...
components:
  schemas:
...
    TransportLoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    TransportLoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/TransportUser"
      required:
        - token
        - user
    TransportUser:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [superadmin, admin, controlador]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tenantId
        - email
        - name
        - role
        - isActive
        - createdAt
        - updatedAt
...